Rename the ast package "node", because it contains nodes
