+++ src/main/java/io/burt/jmespath/node/FlattenArrayNode.java
@@ -1,4 +1,4 @@
-package io.burt.jmespath.ast;
+package io.burt.jmespath.node;
 
 import java.util.List;
 import java.util.LinkedList;
