+++ src/main/java/io/burt/jmespath/node/OrNode.java
@@ -1,9 +1,9 @@
-package io.burt.jmespath.ast;
+package io.burt.jmespath.node;
 
 import io.burt.jmespath.Adapter;
 
-public class AndNode extends OperatorNode {
-  public AndNode(JmesPathNode left, JmesPathNode right) {
+public class OrNode extends OperatorNode {
+  public OrNode(JmesPathNode left, JmesPathNode right) {
     super(left, right);
   }
 
@@ -11,9 +11,9 @@ public class AndNode extends OperatorNode {
   protected <T> T evaluateOne(Adapter<T> adapter, T currentValue) {
     T leftResult = operands()[0].evaluate(adapter, currentValue);
     if (adapter.isTruthy(leftResult)) {
-      return operands()[1].evaluate(adapter, currentValue);
-    } else {
       return leftResult;
+    } else {
+      return operands()[1].evaluate(adapter, currentValue);
     }
   }
 }
