+++ src/test/java/io/burt/jmespath/AstGeneratorTest.java
@@ -10,26 +10,26 @@ import java.util.HashMap;
 import java.util.ArrayList;
 
 import io.burt.jmespath.Query;
-import io.burt.jmespath.ast.AndNode;
-import io.burt.jmespath.ast.ComparisonNode;
-import io.burt.jmespath.ast.CreateArrayNode;
-import io.burt.jmespath.ast.CreateObjectNode;
-import io.burt.jmespath.ast.CurrentNode;
-import io.burt.jmespath.ast.ExpressionReferenceNode;
-import io.burt.jmespath.ast.FlattenArrayNode;
-import io.burt.jmespath.ast.FlattenObjectNode;
-import io.burt.jmespath.ast.ForkNode;
-import io.burt.jmespath.ast.FunctionCallNode;
-import io.burt.jmespath.ast.IndexNode;
-import io.burt.jmespath.ast.JmesPathNode;
-import io.burt.jmespath.ast.JoinNode;
-import io.burt.jmespath.ast.JsonLiteralNode;
-import io.burt.jmespath.ast.NegateNode;
-import io.burt.jmespath.ast.OrNode;
-import io.burt.jmespath.ast.PropertyNode;
-import io.burt.jmespath.ast.SelectionNode;
-import io.burt.jmespath.ast.SliceNode;
-import io.burt.jmespath.ast.StringNode;
+import io.burt.jmespath.node.AndNode;
+import io.burt.jmespath.node.ComparisonNode;
+import io.burt.jmespath.node.CreateArrayNode;
+import io.burt.jmespath.node.CreateObjectNode;
+import io.burt.jmespath.node.CurrentNode;
+import io.burt.jmespath.node.ExpressionReferenceNode;
+import io.burt.jmespath.node.FlattenArrayNode;
+import io.burt.jmespath.node.FlattenObjectNode;
+import io.burt.jmespath.node.ForkNode;
+import io.burt.jmespath.node.FunctionCallNode;
+import io.burt.jmespath.node.IndexNode;
+import io.burt.jmespath.node.JmesPathNode;
+import io.burt.jmespath.node.JoinNode;
+import io.burt.jmespath.node.JsonLiteralNode;
+import io.burt.jmespath.node.NegateNode;
+import io.burt.jmespath.node.OrNode;
+import io.burt.jmespath.node.PropertyNode;
+import io.burt.jmespath.node.SelectionNode;
+import io.burt.jmespath.node.SliceNode;
+import io.burt.jmespath.node.StringNode;
 
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
